address review and small restructure
- move check_file_exist to helpers folder
- move models to it's own folder
- replace create_labelers_file with populate_existing_labelers_file
