+++ apply-labels/src/helpers/check_file_exist.py
@@ -2,5 +2,5 @@ import os
 
 
 def check_file_exist(filepath: str):
-    if not os.path.isfile(filepath):
+    if not os.path.exists(filepath):
         raise Exception("File does not exist")
